refactor: create Select using react-hook-form
